requirements.txt
# Core dependencies
sentence-transformers==2.5.1
huggingface-hub==0.19.4
streamlit==1.32.0
numpy==1.26.4
pandas==2.0.3
python-dotenv==1.0.0

# Document processing
langchain==0.1.12
langchain-community==0.0.27
unstructured==0.10.30
python-magic==0.4.27
python-magic-bin==0.4.14

# Vector store
faiss-cpu==1.7.4

# Local LLM support
ctransformers==0.2.27

# Utilities
tqdm==4.66.1
loguru==0.7.2 